Move metrics components to outlets folder
