+++ src/pages/search/outlets/Species.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Accordion from "pages/search/layout/Accordion";
-import Richness from "pages/search/dashboard/species/Richness";
+import Richness from "pages/search/outlets/species/Richness";
 import { SearchLegacyCTX } from "../hooks/SearchContext";
 
 class Species extends React.Component {
