+++ src/pages/search/outlets/ecosystems/transformData.ts
