+++ src/pages/search/outlets/landscape/CompensationFactorController.ts
