+++ src/pages/search/outlets/landscape/Forest.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import Accordion from "pages/search/layout/Accordion";
-import ForestIntegrity from "pages/search/dashboard/landscape/forest/ForestIntegrity";
-import ForestLossPersistence from "pages/search/dashboard/landscape/forest/ForestLossPersistence";
+import ForestIntegrity from "pages/search/outlets/landscape/forest/ForestIntegrity";
+import ForestLossPersistence from "pages/search/outlets/landscape/forest/ForestLossPersistence";
 import { useSearchLegacyCTX } from "pages/search/hooks/SearchContext";
 import { accordionComponent, componentProps } from "pages/search/types/ui";
