+++ src/pages/search/outlets/landscape/HumanFootprint.tsx
@@ -1,9 +1,9 @@
 import AddIcon from "@mui/icons-material/Add";
 
 import Accordion from "pages/search/layout/Accordion";
-import CurrentFootprint from "pages/search/dashboard/landscape/humanFootprint/CurrentFootprint";
-import PersistenceFooprint from "pages/search/dashboard/landscape/humanFootprint/PersistenceFootprint";
-import TimelineFootprint from "pages/search/dashboard/landscape/humanFootprint/TimelineFootprint";
+import CurrentFootprint from "pages/search/outlets/landscape/humanFootprint/CurrentFootprint";
+import PersistenceFooprint from "pages/search/outlets/landscape/humanFootprint/PersistenceFootprint";
+import TimelineFootprint from "pages/search/outlets/landscape/humanFootprint/TimelineFootprint";
 import { accordionComponent, componentProps } from "pages/search/types/ui";
 
 const HumanFootprint: React.FC<componentProps> = (props) => {
