+++ src/pages/search/outlets/landscape/connectivity/CurrentPAConnectivityController.ts
