+++ src/pages/search/outlets/landscape/connectivity/TimelinePAConnectivityController.ts
