+++ src/pages/search/outlets/landscape/forest/ForestIntegrity.tsx
@@ -23,7 +23,7 @@ import Pie from "pages/search/shared_components/charts/Pie";
 import SmallStackedBar from "pages/search/shared_components/charts/SmallStackedBar";
 import { wrapperMessage } from "pages/search/types/charts";
 import { ShapeLayer } from "pages/search/types/layers";
-import { ForestIntegrityController } from "pages/search/dashboard/landscape/forest/ForestIntegrityController";
+import { ForestIntegrityController } from "pages/search/outlets/landscape/forest/ForestIntegrityController";
 
 type SCIHFCats = `${(typeof SCICats)[number]}-${(typeof HFCats)[number]}`;
 /**
