+++ src/pages/search/outlets/landscape/humanFootprint/CurrentFootprintController.ts
