+++ src/pages/search/outlets/landscape/humanFootprint/PersistenceFootprintController.ts
