+++ src/pages/search/outlets/species/Richness.jsx
